着手object的学习。

+++ ember-guide/.eslintrc.js
@@ -165,7 +165,7 @@ module.exports = {
 		"prefer-const": 0,//首选const
 		"prefer-spread": 0,//首选展开运算
 		"prefer-reflect": 0,//首选Reflect的方法
-		"quotes": ["warn", "double", { "avoidEscape": true, "allowTemplateLiterals": true }],//引号类型 `` "" ''
+		"quotes": ["warn", "single", { "avoidEscape": true, "allowTemplateLiterals": true }],//引号类型 `` "" ''
 		"quote-props": ["error", "consistent", { "keywords": true }],//对象字面量中的属性名是否强制双引号
 		"radix": 2,//parseInt必须指定第二个参数
 		"id-match": 0,//命名检测
